+++ app/scripts/controllers/preferences.js
@@ -19,6 +19,7 @@ class PreferencesController {
    * @property {boolean} store.useBlockie The users preference for blockie identicons within the UI
    * @property {object} store.featureFlags A key-boolean map, where keys refer to features and booleans to whether the
    * user wishes to see that feature
+   * @property {object} store.knownMethodData Contains all data methods known by the user
    * @property {string} store.currentLocale The preferred language locale key
    * @property {string} store.selectedAddress A hex string that matches the currently selected address in the app
    *
@@ -34,10 +35,8 @@ class PreferencesController {
       plugins:[],
       suggestedTokens: {},
       useBlockie: false,
-      featureFlags: {
-        betaUI: true,
-        skipAnnounceBetaUI: true,
-      },
+      featureFlags: {},
+      knownMethodData: {},
       currentLocale: opts.initLangCode,
       identities: {},
       lostIdentities: {},
@@ -46,6 +45,8 @@ class PreferencesController {
       preferences: {
         useNativeCurrencyAsPrimaryCurrency: true,
       },
+      completedOnboarding: false,
+      completedUiMigration: true,
     }, opts.initState)
 
     this.diagnostics = opts.diagnostics
@@ -100,6 +101,18 @@ class PreferencesController {
     this.store.updateState({ suggestedTokens: suggested })
   }
 
+
+  /**
+   * Add new methodData to state, to avoid requesting this information again through Infura
+   *
+   * @param {string} fourBytePrefix Four-byte method signature
+   * @param {string} methodData Corresponding data method
+   */
+  addKnownMethodData (fourBytePrefix, methodData) {
+    const knownMethodData = this.store.getState().knownMethodData
+    knownMethodData[fourBytePrefix] = methodData
+    this.store.updateState({ knownMethodData })
+  }
 
   /**
    * RPC engine middleware for requesting new asset added
@@ -454,6 +467,32 @@ class PreferencesController {
   }
 
   /**
+   * updates custom RPC details
+   *
+   * @param {string} url The RPC url to add to frequentRpcList.
+   * @param {number} chainId Optional chainId of the selected network.
+   * @param {string} ticker   Optional ticker symbol of the selected network.
+   * @param {string} nickname Optional nickname of the selected network.
+   * @returns {Promise<array>} Promise resolving to updated frequentRpcList.
+   *
+   */
+
+
+  updateRpc (newRpcDetails) {
+    const rpcList = this.getFrequentRpcListDetail()
+    const index = rpcList.findIndex((element) => { return element.rpcUrl === newRpcDetails.rpcUrl })
+    if (index > -1) {
+      const rpcDetail = rpcList[index]
+      const updatedRpc = extend(rpcDetail, newRpcDetails)
+      rpcList[index] = updatedRpc
+      this.store.updateState({ frequentRpcListDetail: rpcList })
+    } else {
+      const { rpcUrl, chainId, ticker, nickname } = newRpcDetails
+      return this.addToFrequentRpcList(rpcUrl, chainId, ticker, nickname)
+    }
+    return Promise.resolve(rpcList)
+  }
+  /**
    * Adds custom RPC url to state.
    *
    * @param {string} url The RPC url to add to frequentRpcList.
@@ -562,6 +601,23 @@ class PreferencesController {
    */
   getPreferences () {
     return this.store.getState().preferences
+  }
+
+  /**
+   * Sets the completedOnboarding state to true, indicating that the user has completed the
+   * onboarding process.
+   */
+  completeOnboarding () {
+    this.store.updateState({ completedOnboarding: true })
+    return Promise.resolve(true)
+  }
+
+  /**
+   * Sets the {@code completedUiMigration} state to {@code true}, indicating that the user has completed the UI switch.
+   */
+  completeUiMigration () {
+    this.store.updateState({ completedUiMigration: true })
+    return Promise.resolve(true)
   }
 
   //