+++ app/scripts/migrations/index.js
@@ -41,4 +41,5 @@ module.exports = [
   require('./028'),
   require('./029'),
   require('./030'),
+  require('./031'),
 ]