+++ ui/app/components/send/account-list-item/tests/account-list-item-component.test.js
@@ -121,6 +121,7 @@ describe('AccountListItem Component', function () {
         {
           type: 'PRIMARY',
           value: 'mockBalance',
+          hideTitle: true,
         }
       )
     })