+++ ui/app/components/send/send-content/send-to-row/send-to-row.component.js
@@ -2,7 +2,7 @@ import React, { Component } from 'react'
 import PropTypes from 'prop-types'
 import SendRowWrapper from '../send-row-wrapper/'
 import EnsInput from '../../../ens-input'
-import { getToErrorObject } from './send-to-row.utils.js'
+import { getToErrorObject, getToWarningObject } from './send-to-row.utils.js'
 
 export default class SendToRow extends Component {
 
@@ -10,14 +10,18 @@ export default class SendToRow extends Component {
     closeToDropdown: PropTypes.func,
     hasHexData: PropTypes.bool.isRequired,
     inError: PropTypes.bool,
+    inWarning: PropTypes.bool,
     network: PropTypes.string,
     openToDropdown: PropTypes.func,
+    selectedToken: PropTypes.object,
     to: PropTypes.string,
     toAccounts: PropTypes.array,
     toDropdownOpen: PropTypes.bool,
+    tokens: PropTypes.array,
     updateGas: PropTypes.func,
     updateSendTo: PropTypes.func,
     updateSendToError: PropTypes.func,
+    updateSendToWarning: PropTypes.func,
     scanQrCode: PropTypes.func,
   }
 
@@ -25,11 +29,13 @@ export default class SendToRow extends Component {
     t: PropTypes.func,
   }
 
-  handleToChange (to, nickname = '', toError) {
-    const { hasHexData, updateSendTo, updateSendToError, updateGas } = this.props
+  handleToChange (to, nickname = '', toError, toWarning) {
+    const { hasHexData, updateSendTo, updateSendToError, updateGas, tokens, selectedToken, updateSendToWarning } = this.props
     const toErrorObject = getToErrorObject(to, toError, hasHexData)
+    const toWarningObject = getToWarningObject(to, toWarning, tokens, selectedToken)
     updateSendTo(to, nickname)
     updateSendToError(toErrorObject)
+    updateSendToWarning(toWarningObject)
     if (toErrorObject.to === null) {
       updateGas({ to })
     }
@@ -39,6 +45,7 @@ export default class SendToRow extends Component {
     const {
       closeToDropdown,
       inError,
+      inWarning,
       network,
       openToDropdown,
       to,
@@ -51,7 +58,9 @@ export default class SendToRow extends Component {
         errorType={'to'}
         label={`${this.context.t('to')}: `}
         showError={inError}
-      >
+        showWarning={inWarning}
+        warningType={'to'}
+        >
         <EnsInput
           scanQrCode={_ => this.props.scanQrCode()}
           accounts={toAccounts}
@@ -60,7 +69,7 @@ export default class SendToRow extends Component {
           inError={inError}
           name={'address'}
           network={network}
-          onChange={({ toAddress, nickname, toError }) => this.handleToChange(toAddress, nickname, toError)}
+          onChange={({ toAddress, nickname, toError, toWarning }) => this.handleToChange(toAddress, nickname, toError, toWarning)}
           openDropdown={() => openToDropdown()}
           placeholder={this.context.t('recipientAddress')}
           to={to}