+++ ui/app/components/send/send-content/send-to-row/send-to-row.selectors.js
@@ -1,6 +1,8 @@
 const selectors = {
   getToDropdownOpen,
+  getTokens,
   sendToIsInError,
+  sendToIsInWarning,
 }
 
 module.exports = selectors
@@ -11,4 +13,12 @@ function getToDropdownOpen (state) {
 
 function sendToIsInError (state) {
     return Boolean(state.send.errors.to)
+}
+
+function sendToIsInWarning (state) {
+  return Boolean(state.send.warnings.to)
+}
+
+function getTokens (state) {
+  return state.metamask.tokens
 }