+++ ui/app/reducers/metamask.js
@@ -56,6 +56,8 @@ function reduceMetamask (state, action) {
     preferences: {
       useNativeCurrencyAsPrimaryCurrency: true,
     },
+    completedOnboarding: false,
+    knownMethodData: {},
   }, state.metamask)
 
   switch (action.type) {
@@ -394,6 +396,18 @@ function reduceMetamask (state, action) {
     case actions.UPDATE_PREFERENCES: {
       return extend(metamaskState, {
         preferences: { ...action.payload },
+      })
+    }
+
+    case actions.COMPLETE_ONBOARDING: {
+      return extend(metamaskState, {
+        completedOnboarding: true,
+      })
+    }
+
+    case actions.COMPLETE_UI_MIGRATION: {
+      return extend(metamaskState, {
+        completedUiMigration: true,
       })
     }