+++ CHANGELOG.md
@@ -2,6 +2,30 @@
 
 ## Current Develop Branch
 
+## 5.3.5 Mon Feb 4 2019
+
+- [#6084](https://github.com/MetaMask/metamask-extension/pull/6087): Privacy mode fixes
+
+## 5.3.4 Thu Jan 31 2019
+
+- [#6079](https://github.com/MetaMask/metamask-extension/pull/6079): fix - migration 30
+
+## 5.3.3 Wed Jan 30 2019
+
+- [#6006](https://github.com/MetaMask/metamask-extension/pull/6006): Update privacy notice
+- [#6072](https://github.com/MetaMask/metamask-extension/pull/6072): Improved Spanish translations
+- [#5854](https://github.com/MetaMask/metamask-extension/pull/5854): Add visual indicator when displaying a cached balance.
+- [#6044](https://github.com/MetaMask/metamask-extension/pull/6044): Fix bug that interferred with using multiple custom networks.
+
+## 5.3.2 Mon Jan 28 2019
+
+- [#6021](https://github.com/MetaMask/metamask-extension/pull/6021): Order shapeshift transactions by time within the transactions list
+- [#6052](https://github.com/MetaMask/metamask-extension/pull/6052): Add and use cached method signatures to reduce provider requests
+- [#6048](https://github.com/MetaMask/metamask-extension/pull/6048): Refactor BalanceComponent to jsx
+- [#6026](https://github.com/MetaMask/metamask-extension/pull/6026): Prevent invalid chainIds when adding custom rpcs
+- [#6029](https://github.com/MetaMask/metamask-extension/pull/6029): Fix grammar error in Current Conversion
+- [#6024](https://github.com/MetaMask/metamask-extension/pull/6024): Disable account dropdown on signing screens
+
 ## 5.3.1 Wed Jan 16 2019
 
 - [#5966](https://github.com/MetaMask/metamask-extension/pull/5966): Update Slovenian translation